tool-service/app/requirements.txt
fastapi[standard]
uvicorn
fastmcp
kagglehub
tensorflow[standard]
keras[standard]
numpy
pillow